codegen: move derives and struct_def to types
